Added missing options for repo subcommand
``chef generate repo`` subcommand was missing several options. Added all of the missing options for setting default values, accepting the product license, and working with generator cookbooks. Modified wording of some of the existing flags to match what is displayed by the help message in the latest version of Chef Workstation.